requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==0.61.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
gevent==24.2.1
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
django-debug-toolbar==5.2.0
